\name{RQuantLibExample-package}
\alias{RQuantLibExample-package}
\alias{RQuantLibExample}
\docType{package}
\title{\packageTitle{RQuantLibExample}}
\description{\packageDescription{RQuantLibExample}}
\section{Package Content}{\packageIndices{RQuantLibExample}}
\author{\packageAuthor{RQuantLibExample}}
\section{Maintainer}{\packageMaintainer{RQuantLibExample}}
\keyword{package}
